# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  dto.CreateRoleRequest:
    properties:
      code:
        maxLength: 64
        minLength: 2
        type: string
      dataScope:
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        type: integer
      menuIds:
        items:
          type: integer
        maxItems: 500
        type: array
      name:
        maxLength: 64
        minLength: 2
        type: string
      remark:
        maxLength: 500
        type: string
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
    required:
    - code
    - name
    type: object
...
  dto.StatusRequest:
    properties:
      id:
        type: integer
      status:
        description: |-
          注意：不能用 required —— int8 的零值(0)会被 required 判为「未提供」，
          导致「停用」操作被参数校验直接拦下（返回 400）。
        enum:
        - 0
        - 1
        type: integer
    required:
    - id
    type: object
...
  dto.UpdateRoleRequest:
    properties:
      code:
        maxLength: 64
        minLength: 2
        type: string
      dataScope:
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        type: integer
      id:
        type: integer
      menuIds:
        items:
          type: integer
        maxItems: 500
        type: array
      name:
        maxLength: 64
        minLength: 2
        type: string
      remark:
        maxLength: 500
        type: string
      sort:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
    required:
    - id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Gin-Admin 后台管理系统 API 文档
  title: Gin-Admin API
  version: 1.0.0
paths:
...
  /system/file/{id}:
    delete:
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 删除文件
      tags:
      - 文件
    get:
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取文件详情
      tags:
      - 文件
  /system/file/list:
    get:
      parameters:
      - description: 文件名
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 文件列表
      tags:
      - 文件
  /system/file/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 上传文件
      tags:
      - 文件
...
  /system/role:
    post:
      consumes:
      - application/json
      parameters:
      - description: 角色信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 创建角色
      tags:
      - 角色
    put:
      consumes:
      - application/json
      parameters:
      - description: 角色信息（含 id）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 更新角色
      tags:
      - 角色
  /system/role/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 删除角色
      tags:
      - 角色
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 角色详情
      tags:
      - 角色
  /system/role/all:
    get:
      description: 给用户编辑页的角色下拉用
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 全部角色（不分页）
      tags:
      - 角色
  /system/role/list:
    get:
      parameters:
      - description: 角色名称（模糊）
        in: query
        name: name
        type: string
      - description: 状态
        in: query
        name: status
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 角色列表
      tags:
      - 角色
  /system/role/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: 角色 ID 与目标状态
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.StatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 启用 / 停用角色
      tags:
      - 角色
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: '输入格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"